+++ model/storageusagedto.go
@@ -5,8 +5,8 @@ type StorageUsageDTO struct {
 	FreeSpace       string `json:"freeSpace"`       // Amount of free space.
 	TotalSpace      string `json:"totalSpace"`      // Amount of total space.
 	UsedSpace       string `json:"usedSpace"`       // Amount of used space.
-	FreeSpaceBytes  int    `json:"freeSpaceBytes"`  // The number of bytes of free space.
-	TotalSpaceBytes int    `json:"totalSpaceBytes"` // The number of bytes of total space.
-	UsedSpaceBytes  int    `json:"usedSpaceBytes"`  // The number of bytes of used space.
+	FreeSpaceBytes  int64  `json:"freeSpaceBytes"`  // The number of bytes of free space.
+	TotalSpaceBytes int64  `json:"totalSpaceBytes"` // The number of bytes of total space.
+	UsedSpaceBytes  int64  `json:"usedSpaceBytes"`  // The number of bytes of used space.
 	Utilization     string `json:"utilization"`     // Utilization of this storage location.
 }
